+++ DataAccess/DataObjects/CylinderItems.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DataAccess
+{
+    public class CylinderItems
+    {        
+        public string SupItem { get; set; }
+        public string ItemMisId { get; set; }
+    }
+}
